+++ modules/web-client/src/main/resources/webui/sw.js
@@ -1,9 +1,10 @@
 /* Korus web-client: cache static assets; do NOT show login shell when server is down. */
 "use strict";
 
-var CACHE_NAME = "korus-web-static-v4";
+var CACHE_NAME = "korus-web-static-v5";
 
 var PRECACHE = [
+  "/tailwind.css",
   "/styles.css",
   "/themes.css",
   "/manifest.json",